server/python/requirements.txt
fastapi==0.114.2
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.34
pydantic==2.9.2
pydantic-settings==2.6.0
PyJWT==2.9.0
python-multipart==0.0.12
websockets==13.1
